backend/scripts/pyproject.toml
[project]
name = "py-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.18"
dependencies = [
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.13.2",
    "selenium>=4.35.0",
    "undetected-chromedriver>=3.5.5",
]
